Upgrade Go, kinda (but also still test against older Go)
With Go 1.19 comes with a change in how `gofmt` formats comments.

I had to upgrade golangci-lint to 1.48 in order to have compatibility for
Go 1.19; with this I had to fuss with the config a bit, and make a
whitespace change in envconfig_types.go.  I sorta cheated and marked
"nosnakecase" as deprecated, even though it isn't yet deprecated in
golangci-lint 1.48.0, it is deprecated in some later version.

Signed-off-by: Luke Shumaker <lukeshu@datawire.io>