Fix glitch in comma-split-empty giving one element slice on empty input
A `strings.Split("", ",")` annoyingly produces a one element slice with
an empty string in it. This is somewhat counterintuitive and doesn't
align well with other types that produces a zero value for `default=`.

This commit ensures that an empty string instead yields an empty slice.

A consequence of this change is that it's now impossible to get a slice
containing one empty string. I think we can live with that.